Add files via upload
So the project should run without any issues. I unit tested all the components and tried to separate and objectify as much as possible. There is a UML diagram showing what is connected. 

I have a couple classes in there defined as chipID and such but they were never implemented. I was attempting to add-in for bonus points, a recyclable chip system for the Casino. Since a regular Casino has uniquely identifiable chips I was hoping to add that in as a clever gimmick and to use chips as objects but I didn't have enough time when the project was assigned.